organize config-install tests and new one more complete
